+++ src/components/Blocks/Button/schema.js
@@ -2,9 +2,9 @@ import { defineMessages } from 'react-intl';
 import { defaultStylingSchema } from '../schema';
 
 const messages = defineMessages({
-  Align: {
-    id: 'Alignment',
-    defaultMessage: 'Alignment',
+  BlockWidth: {
+    id: 'Block Width',
+    defaultMessage: 'Block Width',
   },
 });
 
@@ -18,7 +18,7 @@ export const ButtonStylingSchema = ({ schema, formData, intl }) => {
 
   schema.properties.styles.schema.properties.buttonAlign = {
     widget: 'align',
-    title: intl.formatMessage(messages.Align),
+    title: intl.formatMessage(messages.BlockWidth),
     actions: ['center', 'wide'],
   };